extract core in leptos examples
